+++ src/types.ts
@@ -1,5 +1,5 @@
 import type {Cursor}                         from "./cursor";
-import type {Erase}                          from "./erase";
+import type {Delete, Erase, Insert}          from "./ins-del";
 import type {Features}                       from "./features";
 import type {PadEnd, PadStart}               from "./pad";
 import type {Scroll}                         from "./scroll";
@@ -12,7 +12,9 @@ import type {Terminal}                       from "./terminal";
 
 export type Ansi = SGR & {
 	cursor   : Cursor;
+	delete   : Delete;
 	erase    : Erase;
+	insert   : Insert;
 	scroll   : Scroll;
 	terminal : Terminal;